LayoutTests/js/dom/function-decompilation-operators-expected.txt

This test checks toString() round-trip decompilation for binary and unary operators.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS decompiledFunction is 'function () {  x + + y;}'
PASS decompiledFunction is 'function () {  x + - y;}'
PASS decompiledFunction is 'function () {  x - + y;}'
PASS decompiledFunction is 'function () {  x - - y;}'
PASS decompiledFunction is 'function () {  x * + y;}'
PASS decompiledFunction is 'function () {  x * - y;}'
PASS decompiledFunction is 'function () {  x / + y;}'
PASS decompiledFunction is 'function () {  x / - y;}'
PASS decompiledFunction is 'function () {  x % + y;}'
PASS decompiledFunction is 'function () {  x % - y;}'
PASS decompiledFunction is 'function () {  x++ + y;}'
PASS decompiledFunction is 'function () {  x++ - y;}'
PASS decompiledFunction is 'function () {  x++ * y;}'
PASS decompiledFunction is 'function () {  x++ / y;}'
PASS decompiledFunction is 'function () {  x-- + y;}'
PASS decompiledFunction is 'function () {  x-- - y;}'
PASS decompiledFunction is 'function () {  x-- * y;}'
PASS decompiledFunction is 'function () {  x-- / y;}'
PASS decompiledFunction is 'function () {  x + ++y;}'
PASS decompiledFunction is 'function () {  x - ++y;}'
PASS decompiledFunction is 'function () {  x * ++y;}'
PASS decompiledFunction is 'function () {  x / ++y;}'
PASS decompiledFunction is 'function () {  x + --y;}'
PASS decompiledFunction is 'function () {  x - --y;}'
PASS decompiledFunction is 'function () {  x * --y;}'
PASS decompiledFunction is 'function () {  x / --y;}'
PASS decompiledFunction is 'function () {  x++ + ++y;}'
PASS decompiledFunction is 'function () {  x++ - ++y;}'
PASS decompiledFunction is 'function () {  x++ * ++y;}'
PASS decompiledFunction is 'function () {  x++ / ++y;}'
PASS decompiledFunction is 'function () {  x-- + ++y;}'
PASS decompiledFunction is 'function () {  x-- - ++y;}'
PASS decompiledFunction is 'function () {  x-- * ++y;}'
PASS decompiledFunction is 'function () {  x-- / ++y;}'
PASS decompiledFunction is 'function () {  x++ + --y;}'
PASS decompiledFunction is 'function () {  x++ - --y;}'
PASS decompiledFunction is 'function () {  x++ * --y;}'
PASS decompiledFunction is 'function () {  x++ / --y;}'
PASS decompiledFunction is 'function () {  x-- + --y;}'
PASS decompiledFunction is 'function () {  x-- - --y;}'
PASS decompiledFunction is 'function () {  x-- * --y;}'
PASS decompiledFunction is 'function () {  x-- / --y;}'
PASS decompiledFunction is 'function () {  + + x;}'
PASS decompiledFunction is 'function () {  + - x;}'
PASS decompiledFunction is 'function () {  - + x;}'
PASS decompiledFunction is 'function () {  - - x;}'
PASS decompiledFunction is 'function () {  1;}'
PASS decompiledFunction is 'function () {  -1;}'
PASS decompiledFunction is 'function () {  - -1;}'
PASS decompiledFunction is 'function () {  - - 0;}'
PASS decompiledFunction is 'function () {  - - NaN;}'
PASS successfullyParsed is true

TEST COMPLETE
